+++ packages/bridge-status-controller/jest.config.js
@@ -18,8 +18,20 @@ module.exports = merge(baseConfig, {
 
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
+    './src/bridge-status-controller.ts': {
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100,
+    },
+    './src/bridge-status-controller.intent.ts': {
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100,
+    },
     global: {
-      branches: 93.31,
+      branches: 96.5,
       functions: 100,
       lines: 100,
       statements: 100,