+++ packages/bridge-status-controller/src/bridge-status-controller.intent.ts
@@ -1,4 +1,4 @@
-import { StatusTypes } from '@metamask/bridge-controller';
+import { BridgeClientId, StatusTypes } from '@metamask/bridge-controller';
 import type { TransactionController } from '@metamask/transaction-controller';
 import { TransactionMeta } from '@metamask/transaction-controller';
 
@@ -12,7 +12,7 @@ import {
   IntentSubmissionParams,
   translateIntentOrderToBridgeStatus,
 } from './utils/intent-api';
-import { IntentOrder, IntentOrderStatus } from './utils/validators';
+import { IntentStatusResponse, IntentOrderStatus } from './utils/validators';
 
 type IntentStatuses = {
   orderStatus: IntentOrderStatus;
@@ -59,14 +59,14 @@ export class IntentManager {
 
   #setIntentStatuses(
     bridgeTxMetaId: string,
-    order: IntentOrder,
+    order: IntentStatusResponse,
     srcChainId: number,
     txHash: string,
   ): IntentStatuses {
     const bridgeStatus = translateIntentOrderToBridgeStatus(
       order,
       srcChainId,
-      txHash.toString(),
+      txHash,
     );
     const intentStatuses: IntentStatuses = {
       orderStatus: order.status,
@@ -87,21 +87,16 @@ export class IntentManager {
 
   getIntentTransactionStatus = async (
     bridgeTxMetaId: string,
-    historyItem: BridgeHistoryItem,
-    clientId: string,
+    srcChainId: number,
+    protocol: string,
+    clientId: BridgeClientId,
+    txHash: string = '',
   ): Promise<IntentStatuses | undefined> => {
-    const {
-      status: statusObj,
-      quote: { srcChainId, intent },
-    } = historyItem;
-    const txHash = statusObj?.srcChain?.txHash ?? '';
-    const protocol = intent?.protocol ?? '';
-
     try {
       const orderStatus = await this.intentApi.getOrderStatus(
         bridgeTxMetaId,
         protocol,
-        srcChainId.toString(),
+        srcChainId,
         clientId,
       );
 
@@ -213,8 +208,8 @@ export class IntentManager {
    */
   submitIntent = async (
     submissionParams: IntentSubmissionParams,
-    clientId: string,
-  ): Promise<IntentOrder> => {
+    clientId: BridgeClientId,
+  ): Promise<IntentStatusResponse> => {
     return this.intentApi.submitIntent(submissionParams, clientId);
   };
 }