+++ packages/profile-metrics-controller/src/ProfileMetricsService.ts
@@ -214,6 +214,11 @@ export class ProfileMetricsService {
           metametrics_id: data.metametricsId,
           accounts: data.accounts,
         }),
+        // The auth API is stateless (no cookies used)
+        // prevent marketing cookies scoped to
+        // .metamask.io from being forwarded to api which
+        // causes 431 Request Header Fields Too Large errors.
+        credentials: 'omit',
       });
       if (!localResponse.ok) {
         throw new HttpError(